vmm: remap guest mem periodically
In order to work around the double-counting issue witnessed on macOS,
we need to remap host memory pages on top of themselves. This issue
happens whenever a page is read or written by both sides. As virtio
devices need to read / write pages that will be accessed by the guest
too, remap all the memory every 5 seconds to ensure that pages counted
twice are deducted from the host phys footprint.

Signed-off-by: Albin Kerouanton <albinker@gmail.com>